% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPlots.R
\name{average_status}
\alias{average_status}
\title{Calculate the average methylation/accessibility status across all reads.}
\usage{
average_status(orderObject, window_length = 1, plotAverages = FALSE, ...)
}
\arguments{
\item{orderObject}{An object of class \code{orderObject}}

\item{window_length}{Length of the window to be used to compute a moving average.}

\item{plotAverages}{Logical, indicates whether to generate a line plot of average status.}

\item{...}{Addition parameters used by the \code{plot} function.}
}
\description{
Calculate the average methylation/accessibility status across all reads.
}
