% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioHelpers.R
\name{check_package}
\alias{check_package}
\title{Check if a package is installed}
\usage{
check_package(package)
}
\arguments{
\item{package}{Name of the package.}
}
\value{
Message is returned if package not installed.
}
\description{
Check if a package is installed
}
