% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr19_example_HCG}
\alias{chr19_example_HCG}
\title{Example HCG file for single-cell experiment}
\format{
An object of class \code{list} of length 61.
}
\usage{
data(chr19_example_HCG)
}
\description{
This data GSE109262, and has been pre-processed by methylscaper
}
\keyword{datasets}
