% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day7}
\alias{day7}
\title{Example single-molecule GCH and HCG matrices}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(day7)
}
\description{
This was made after running methylscaper::runAlign(fasta=reads.day7, 
ref=ref.seq)
}
\keyword{datasets}
