% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hum.bm}
\alias{hum.bm}
\title{Human gene symbols and positions}
\format{
An object of class \code{data.frame} with 60573 rows and 4 columns.
}
\usage{
data(hum.bm)
}
\description{
library(biomaRt) #v2.44.4
ensembl <- useMart("ensembl")
ensembl <- useDataset("hsapiens_gene_ensembl",mart=ensembl)
my_chr <- c(1:22, 'M', 'X', 'Y')
hum.bm <- getBM(attributes=c('chromosome_name', 'start_position', 'end_position', 'hgnc_symbol'),
            filters = 'chromosome_name',
            values = my_chr,
            mart=ensembl)
}
\keyword{datasets}
