% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_bm}
\alias{human_bm}
\title{Human gene symbols and positions}
\format{
An object of class \code{data.frame} with 60580 rows and 4 columns.
}
\usage{
data(human_bm)
}
\description{
library(biomaRt) #v2.44.4
ensembl <- useMart("ensembl") # GRCh38
ensembl <- useDataset("hsapiens_gene_ensembl",mart=ensembl)
my_chr <- c(1:22, 'M', 'X', 'Y')
human_bm <- getBM(attributes=c('chromosome_name', 'start_position', 'end_position', 'hgnc_symbol'),
            filters = 'chromosome_name',
            values = my_chr,
            mart=ensembl)
}
\keyword{datasets}
