% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPlots.R
\name{methyl_proportion_cell}
\alias{methyl_proportion_cell}
\title{Calculate the proportion of methylated bases for each cell/molecule}
\usage{
methyl_proportion_cell(orderObject, color = "yellow", makePlot = TRUE, ...)
}
\arguments{
\item{orderObject}{An object of class \code{orderObject}}

\item{color}{Indicates which data set to compute proportions for.
This should be 'red' or 'hcg' for endogenous methylation; 'yellow' or
'gch' for accessibility.}

\item{makePlot}{Indicates whether to plot a histogram of the proportions 
across all reads.}

\item{...}{Additional parameters used by the \code{hist} function.}
}
\value{
The proportion of methylated (endogenous (yellow) 
     or accessible) bases for each cell/molecule. Output is vector
     with length the numbner of cells/molecules and contains a proportion.
}
\description{
Calculate the proportion of methylated bases for each cell/molecule
}
\examples{
 
data(day7)

orderObj <- initialOrder(day7$gch, day7$hcg, Method = "PCA")
methyl_proportion_cell(orderObj, makePlot = TRUE)
}
