% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{methylscaper}
\alias{methylscaper}
\title{methylscaper}
\usage{
methylscaper()
}
\value{
This starts up the shiny app interface for methylscaper.
}
\description{
Runs the methylscaper Shiny app.
}
\examples{

# methylscaper()
}
