% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse.bm}
\alias{mouse.bm}
\title{Mouse gene symbols and positions}
\format{
An object of class \code{data.frame} with 53610 rows and 4 columns.
}
\usage{
data(mouse.bm)
}
\description{
library(biomaRt) #v2.44.4
ensembl <- useMart("ensembl")
ensembl <- useDataset("mmusculus_gene_ensembl",mart=ensembl)
my_chr <- c(1:19, 'M', 'X', 'Y')
mouse.bm <- getBM(attributes=c('chromosome_name', 'start_position', 'end_position', 'mgi_symbol'),
            filters = 'chromosome_name',
            values = my_chr,
            mart=ensembl)
}
\keyword{datasets}
