% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPlots.R
\name{proportion_color}
\alias{proportion_color}
\title{Calculate the proportion of methylated bases for the GCH and HCG data sets.}
\usage{
proportion_color(orderObject, color = "YELLOW", plotHistogram = FALSE)
}
\arguments{
\item{orderObject}{An object of class \code{orderObject}}

\item{color}{Indicates which data set to compute proportions for}
}
\description{
Calculate the proportion of methylated bases for the GCH and HCG data sets.
}
