% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioHelpers.R
\name{readMethylationData}
\alias{readMethylationData}
\title{Reading methylation data files}
\usage{
readMethylationData(filepath, ...)
}
\arguments{
\item{filepath}{A string of the file path.}

\item{...}{Additional parameters passed to \code{read.table}.}
}
\description{
Reading methylation data files
}
