% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reads_day7}
\alias{reads_day7}
\title{Example reads from single-molecule experiment}
\format{
An object of class \code{list} of length 293.
}
\usage{
data(reads_day7)
}
\description{
This dataset was loaded into R using seqinr::read.fasta
}
\keyword{datasets}
