% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref.seq}
\alias{ref.seq}
\title{Example reference sequence to align reads to from a 
single-molecule experiment}
\format{
An object of class \code{list} of length 1.
}
\usage{
ref.seq
}
\description{
This dataset was loaded into R using seqinr::read.fasta
}
\keyword{datasets}
