% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinement.R
\name{refineFunction}
\alias{refineFunction}
\title{Refinement step}
\usage{
refineFunction(
  orderObject,
  refineStart,
  refineEnd,
  reverse = FALSE,
  Method = "HC_average"
)
}
\arguments{
\item{orderObject}{An object of class \code{orderObject}, generated with the \code{initialOrder} function.}

\item{refineStart}{The index of the first sample (row) used in the refinement.}

\item{refineEnd}{The index of the last sample used in the refinement.}

\item{reverse}{Logical, indicates whether to reverse the refined ordering.}

\item{Method}{The seriation method used to perform the refinement.}
}
\value{
The new complete ordering with the refinement applied.
}
\description{
Reorders a subset of the methylation data.
}
