% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{singlecell_subset}
\alias{singlecell_subset}
\title{Example preprocessed single-cell experiment subset}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(singlecell_subset)
}
\description{
This data is from GSE109262, and has been pre-processed by methylscaper
It contains a small subaet of chromosome 19 region from 8956841bp - 8976841bp.
}
\keyword{datasets}
