% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{singlemolecule_example}
\alias{singlemolecule_example}
\title{Example preprocessed single-molecule experiment}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(singlemolecule_example)
}
\description{
The RDS in ext data was made specifically with the command:
singlemolecule_example <- methylscaper::runAlign(fasta=reads_sm, 
ref=ref_seq)
saveRDS(singlemolecule_example, file="methylscaper/inst/ext/singlemolecule_example.rds", compress = 'xz')
A version is also saved as RData used running examples in the man pages.
save(singlemolecule_example, file="methylscaper/data/singlemolecule_example.RData", compress = 'xz')
}
\keyword{datasets}
