% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioHelpers.R
\name{writeMethylationData}
\alias{writeMethylationData}
\title{Writing methylation data matrices}
\usage{
writeMethylationData(dat, filepath, ...)
}
\arguments{
\item{dat}{A methylation data matrix, output from the
\code{runAlign} function.}

\item{filepath}{A string of the file path.}

\item{...}{Additional parameters passed to \code{write.table}.}
}
\value{
This function writes the GCH and HCG matrices as a csv file
 by default, although a tab separated file is also valid and can be 
 specified as .txt or .tsv.
}
\description{
Writing methylation data matrices
}
\examples{

data(day7)
# writeMethylationData(day7$gch, "GCH.csv")

}
