% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day7}
\alias{day7}
\title{Example methylation data}
\format{An object of class \code{list} of length 2.}
\usage{
day7
}
\description{
An example methylation dataset, which includes a GCH and an HCG matrix.
}
\keyword{datasets}
