% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialOrder.R
\name{initialOrder}
\alias{initialOrder}
\title{Ordering of methylation data}
\usage{
initialOrder(
  input.GCH,
  input.HCG,
  Method = "PCA",
  weightStart = NULL,
  weightEnd = NULL,
  weightFeature = "red",
  reverse = F
)
}
\arguments{
\item{input.GCH}{The GCH input data.}

\item{input.HCG}{The HCG input data.}

\item{Method}{Indicates the seriation method to use. The default option is "PCA", which orders the data using the
first principal component. Any seriation method provided in the \code{seriation} package is valid.}

\item{weightStart}{Index of the first column used in the weighted seriation.}

\item{weightEnd}{Index of the last column used in the weighted seriation.}

\item{weightFeature}{Indicates whether to weight the GCH or HCG data.}

\item{reverse}{Logical, indicates whether to reverse the ordering.}
}
\value{
An object of class \code{orderObject}, which contains the generated ordering and the cleaned data matrix.
}
\description{
Orders methylation data using a given seriation method. This function can also perform a weighted seriation if the method is set to "PCA".
}
