% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessSingleCell.R
\name{prepSC}
\alias{prepSC}
\title{Process single-cell data}
\usage{
prepSC(gc_seq_data, cg_seq_data, startPos, endPos, updateProgress = NULL)
}
\arguments{
\item{gc_seq_data}{GC accessibility data, in the form of a list of dataframes.}

\item{cg_seq_data}{CG methylation data, in the form of a list of dataframes.}

\item{startPos}{The index of the first position to include in the visualization.}

\item{endPos}{The index of the final position to include in the visualization.}

\item{updateProgress}{A function for generating progress bars in the Shiny app. Should be left NULL otherwise.}
}
\description{
This function generates representation methylation-state matrices from single-cell methylation data (for example, as generated by the Bismarck software program). See the vignette for a more thorough explanation of each parameter.
}
