% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioHelpers.R
\name{readMethylationData}
\alias{readMethylationData}
\title{Reading methylation data files}
\usage{
readMethylationData(filepath, ...)
}
\arguments{
\item{filepath}{A string of the file path.}

\item{...}{Additional parameters passed to \code{read.table}.}
}
\value{
This function reads in the GCH and HCG matrices if
 they have been previously saved within the Shiny app, or
 via the writeMethylationData function.
}
\description{
Reading methylation data files
}
\examples{

# my.gchdata <- readMethylationData("GCH.csv")
# my.hcgdata <- readMethylationData("HCG.csv")
}
