% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessSingleMolecule.R
\name{runAlign}
\alias{runAlign}
\title{runAlign}
\usage{
runAlign(
  ref,
  fasta,
  fastaSubset = (seq_len(fasta)),
  multicoreParam = NULL,
  updateProgress = NULL,
  logFile = NULL
)
}
\arguments{
\item{ref}{A reference sequence}

\item{fasta}{A list of sequences}

\item{fastaSubset}{A vector of indices indicating which 
sequences to process.}

\item{multicoreParam}{A MulticoreParam object, 
used to align sequences in parallel.}

\item{updateProgress}{Used to add a progress bar to the Shiny app. 
Should not be used otherwise.}

\item{logFile}{String indicating where to save a log of the alignment 
process. If left NULL, no log is saved.}
}
\description{
Runs the preprocessing methods for single-molecule data on sequences.
}
