% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessSingleMolecule.R
\name{runAlign}
\alias{runAlign}
\title{runAlign}
\usage{
runAlign(
  ref,
  fasta,
  fasta.subset = (1:length(fasta)),
  multicoreParam = NULL,
  updateProgress = NULL,
  log.file = NULL
)
}
\arguments{
\item{ref}{A reference sequence}

\item{fasta}{A list of sequences}

\item{fasta.subset}{A vector of indices indicating which sequences to process.}

\item{multicoreParam}{A MulticoreParam object, used to align sequences in parallel.}

\item{updateProgress}{Used to add a progress bar to the Shiny app. Should not be used otherwise.}

\item{log.file}{String indicating where to save a log of the alignment process. If left NULL, no log is saved.}
}
\description{
Runs the preprocessing methods for single-molecule data on sequences.
}
