% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{singlecell_subset}
\alias{singlecell_subset}
\title{Example preprocessed single-cell experiment subset}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(singlecell_subset)
}
\description{
This data is from GSE109262, and has been pre-processed by methylscaper
It contains a small subset of chromosome 19 region from 8947041bp - 8987041bp.
The RDS in ext data was made specifically with the two commands:
singlecell_subset <- subsetSC("~/Downloads/GSE109262_RAW/", chromosome="19", 
   startPos = 8967041-20000, endPos = 8967041+20000, updateProgress = NULL)
saveRDS(singlecell_subset, file="methylscaper/inst/ext/singlecell_subset.rds", compress = 'xz')
}
\keyword{datasets}
