% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{singlemolecule_example}
\alias{singlemolecule_example}
\title{Example single-molecule GCH and HCG matrices}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(singlemolecule_example)
}
\description{
This was made after running methylscaper::runAlign(fasta=reads_sm, 
ref=ref_seq)
}
\keyword{datasets}
