% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioHelpers.R
\name{writeMethylationData}
\alias{writeMethylationData}
\title{Writing methylation data matrices}
\usage{
writeMethylationData(dat, filepath, ...)
}
\arguments{
\item{dat}{A methylation data matrix, output from the \code{runAlign} function.}

\item{filepath}{A string of the file path.}

\item{...}{Additional parameters passed to \code{write.table}.}
}
\description{
Writing methylation data matrices
}
